requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scipy>=1.10.0
nltk>=3.8.0
scikit-learn>=1.3.0
tqdm>=4.65.0
requests>=2.31.0
jsonschema>=4.17.0
colorama>=0.4.6
tabulate>=0.9.0
